Promote position discipline and terrain valuation to Princess
Every bot now prices what CASPAR proved out: the hold credit for
keeping a positive-exchange position, the attacker movement modifier
on its own fire, the cover at its destination against unmoved
enemies, the heat-sink sustained-output boost, posture resolution
with chat announcements, and the flat closing charge for a defender
in contact. The mechanisms move from the CASPAR ranker into
BasicPathRanker; the seams stay, now implemented instead of neutral.

CASPAR keeps only its own doctrine on top: mutual-support cohesion,
the set-friend cover bonus, turns-at-own-speed closing tempo, and
the water posture charge. Its overrides of the promoted mechanisms
are deleted - both bots share one implementation.

The defender-tempo port: Princess keeps her stock per-hex aggression
gradient, gated to zero under a resolved DEFEND with the enemy
inside contact range, mirroring the CASPAR rule.

Tests: the neutrality pins invert into Princess-parity pins (same
discounts, same hold credit, same defender gate); the legacy
evaluateUnmovedEnemy/rankPath tests hold the new pricing neutral so
they keep pinning the LOS/kick/utility arithmetic they were written
for. TSV: every bot now records combatPosture, positionQuality, and
positionHoldMod.
